Changed property variable names, minor bug fix
Changed the names of property variables in DimDoorsConfig to follow
better variable naming conventions. Removed a duplicate config.save()
call. Noticed that crafting Stabilized Rift Signatures can't be disabled
in the config file, will be adding that option in the next commit.
General note: we need a better way of handling item configuration.
Metaprogramming in Java would be a godsend.