Fixed mcmod.info
Fixed problems in mcmod.info. Changed modid and version to match the
@Mod Java annotations. If these annotations don't match the information
in the file, the file is ignored! So previous versions did not have
their mod info or logo appear under the mod menu in Minecraft.

The credits line is too long, unfortunately. It runs off my screen and
there is no attempt to wrap it back. Our best bet would be to place the
credits in the description instead. I tried placing newlines but they're
not rendered properly on the client.