Fixed Mistakes in Dungeon Tutorial
Fixed mistakes in the text file that explains how to add new dungeons.
The example provided in the file was still wrong since the last time I
posted about it. >_< I also noticed that since we only check if the file
is present, and not if it's different from the one in our latest
release, old copies with mistakes will stick around anyway. I'll tweak
our code for that once other issues have been resolved.